Load dicom deidentifier js as module (#4471)
Load the dicom deidentifier js as an ES module. 
Run jest with `--experimental-vm-modules` to import from the module in
the tests. This feature is experimental, but is running without issues.

Additionally the `uidMap` is now defined as a global and not recreated
if it already exists, making it available across multiple widgets on a
page.

This resolves the following errors (Shown in console on redisplay of
form after validation errors)
- Syntax error: '_uidMap' is already defined. 
- Syntax error: 'DicomDeidentifierPlugin' is already defined.

Closes https://github.com/DIAGNijmegen/rse-grand-challenge/issues/4470

Alternative to #4472