Create ImageSourceChoiceField (#4475)
A form field for selecting the data source of an image. 

If instantiated with a current socket value, it has `IMAGE_SELECTED` as
the first option and it displays the current socket value title. It will
not have the `UNDEFINED` option in this case (we don't support removing
image socket values).
If instantiated without a current socket value, it will not have the
`IMAGE_SELECTED` option and will instead show the `UNDEFINED` option. If
the `UNDEFINED` option is submitted it will return a validation error
that the field is required.

The field is not yet used anywhere. 

Related to https://github.com/DIAGNijmegen/rse-roadmap/issues/435